requirements.txt
anywidget==0.9.21
asttokens==3.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
idna==3.11
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jupyter_bbox_widget==0.6.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
prompt_toolkit==3.0.52
psygnal==0.15.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
sam3==0.0.1
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
supervision==0.27.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.2
wcwidth==0.2.14
widgetsnbextension==4.0.15
